+++ R/utils.R
@@ -157,11 +157,6 @@ strip_internal_calls <- function(x, package) {
   }
 }
 
-sort_ci <- function(x) {
-  local_collate("C")
-  x[order(tolower(x), x)]
-}
-
 dev_packages <- function() {
   packages <- vapply(
     loadedNamespaces(),
@@ -204,8 +199,6 @@ delayed_assign <- function(
   inject(delayedAssign(x, !!value, eval.env, assign.env))
 }
 
-last <- function(x) utils::tail(x, n = 1L)
-
 single_quote <- function(x) {
   encodeString(x, quote = "'")
 }
@@ -233,10 +226,6 @@ cat_line <- function(...) {
 
 is_rstudio <- function() {
   is_string(.Platform$GUI, "RStudio")
-}
-
-rstudioapi_available <- function() {
-  is_installed("rstudioapi") && rstudioapi::isAvailable()
 }
 
 is_windows <- function() {